# openapi.yaml
openapi: 3.0.1
info:
  title: Space NeoHoods API
  version: "2.0"
  description: |
    # Summary
    Space NeoHoods API
servers:
  - url: https://space.neohoods.com
    description: Space neohoods API
paths:
...
  /hub/profile:
    get:
      operationId: getProfile
      summary: Get the current user's profile
      tags:
        - profile-hub-api
      responses:
        "200":
          description: get profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    put:
      operationId: updateProfile
      summary: Update the current user's profile
      tags:
        - profile-hub-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
        "404":
          description: User not found
  /admin/users:
    get:
      operationId: getUsers
      summary: Retrieve all users
      tags:
        - users-admin-api
      responses:
        "200":
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWithStats"
    post:
      operationId: saveUser
      summary: Save a user
      tags:
        - users-admin-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserWithStats"
      responses:
        "201":
          description: User saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithStats"
  /admin/users/{user-id}:
    parameters:
      - name: user-id
        in: path
        description: User ID
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getUser
      summary: Retrieve a specific user by ID
      tags:
        - users-admin-api
      responses:
        "200":
          description: A single user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithStats"
  /admin/users/{user-id}/password:
    parameters:
      - name: user-id
        in: path
        description: User ID
        required: true
        schema:
          type: string
          format: uuid
    put:
      operationId: setUserPassword
      summary: Set a user's password
      tags:
        - users-admin-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  example: "newSecurePassword"
              required:
                - newPassword
      responses:
        "204":
          description: Password updated successfully
...
components:
...
  securitySchemes:
    BearerAuthOAuth:
      description: Bearer Token.
      type: http
      scheme: bearer
...
  schemas:
    User:
      type: object
      required:
        - id
        - login
        - username
        - firstName
        - lastName
        - roles
      properties:
        id:
          type: string
          format: uuid
        disabled:
          type: boolean
        isEmailVerified:
          type: boolean
        username:
          type: string
        avatarUrl:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        flatNumber:
          type: string
          example: A701
        roles:
          type: array
          items:
            type: string
            example: "admin"
        streetAddress:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        preferredLanguage:
          type: string
          example: "en"
    UserWithStats:
      type: object
      required:
        - id
        - username
        - firstName
        - lastName
        - email
        - borrowedItems
        - returnedLate
        - successRate
        - disabled
        - isEmailVerified
        - streetAddress
        - city
        - postalCode
        - country
        - roles
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        borrowedItems:
          type: integer
        returnedLate:
          type: integer
        successRate:
          type: number
        disabled:
          type: boolean
        isEmailVerified:
          type: boolean
        flatNumber:
          type: string
        streetAddress:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        roles:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
        preferredLanguage:
          type: string
...